merge

+++ webSite/questionnaires/management_questionnaire.php
@@ -58,8 +58,11 @@ if (isset($_POST['action'])) {
 			$stmt->execute();
 
 			$stmt = $conn->prepare("INSERT INTO questions (question, answer, key_questionnaires)
-				SELECT :question, :answer, id FROM questionnaires WHERE title like :title");
-			$stmt->bindParam(':title', $title);
+				SELECT :question, :answer, q.id 
+				FROM questionnaires q 
+				INNER JOIN firsturl fu ON q.key_first_url = fu.id 
+				WHERE url like :url");
+			$stmt->bindParam(':url', $url);
 			$stmt->bindParam(':question', $question);
 			$stmt->bindParam(':answer', $answer);
 
@@ -72,7 +75,7 @@ if (isset($_POST['action'])) {
 		}
 		elseif ($_POST['action'] == 'get_questions_to_edit') {
 			if (isset($_POST['id'])) {
-				$stmt = $conn->prepare("SELECT qtn.title, qtn.type, qtn.link, q.question, q.id
+				$stmt = $conn->prepare("SELECT qtn.title, qtn.type, qtn.link, q.question, q.answer, q.id, q.key_questionnaires
 					FROM questionnaires qtn 
 					INNER JOIN questions q ON qtn.id = q.key_questionnaires
 					WHERE qtn.key_first_url = (SELECT id FROM firsturl WHERE url LIKE :url)");
@@ -84,15 +87,71 @@ if (isset($_POST['action'])) {
 			}
 		}
 		elseif ($_POST['action'] == 'edit') {
-			# code...
+			if (isset($_POST['id_questionnaire'])) {
+				header("Location: http://163.172.59.102/webSite/questionnaires/questionnaire.html");
+
+				$stmt = $conn->prepare("UPDATE questionnaires 
+					SET title = :title, type = :type, link = :link 
+					WHERE id = :id");
+				$stmt->bindParam(':title', $title);
+				$stmt->bindParam(':type', $type);
+				$stmt->bindParam(':link', $link);
+				$stmt->bindParam(':id', $id);
+				$title = $_POST['title'];
+				$type = ($_POST['type'] == 'article') ? 0 : 1;
+				$link = $_POST['data'];
+				$id = $_POST['id_questionnaire'];
+				$stmt->execute();
+
+				$stmt = $conn->prepare("UPDATE questions 
+					SET question = :question, answer = :answer
+					WHERE id = :id");
+				$stmt->bindParam(':id', $id);
+				$stmt->bindParam(':question', $question);
+				$stmt->bindParam(':answer', $answer);
+
+				foreach ($_POST['q'] as $key => $value) {
+					$id = $key;
+					$question = $value;
+					$answer = strtoupper(stripVN($_POST['a'][$key]));
+
+					$stmt->execute();
+				}
+
+				$stmt = $conn->prepare("INSERT INTO questions (question, answer, key_questionnaires)
+					VALUES(:question, :answer, :id)");
+				$stmt->bindParam(':question', $question);
+				$stmt->bindParam(':answer', $answer);
+				$stmt->bindParam(':id', $id);
+				$id = $_POST['id_questionnaire'];
+
+				foreach ($_POST['nq'] as $key => $value) {
+					$question = $value;
+					$answer = strtoupper(stripVN($_POST['na'][$key]));
+
+					$stmt->execute();
+				}
+			}
 		}
 		elseif ($_POST['action'] == 'delete') {
-			$stmt = $conn->prepare("DELETE FROM firsturl WHERE url like :url");
-			$stmt->bindParam(':url', $url);
-			$url = $_POST['url'];
-			$stmt->execute();
+			if (isset($_POST['url'])) {
+				$stmt = $conn->prepare("DELETE FROM firsturl WHERE url like :url");
+				$stmt->bindParam(':url', $url);
+				$url = $_POST['url'];
+				$stmt->execute();
 
-			echo "The questionnaire has been deleted";
+				echo "The questionnaire has been deleted";
+			}
+		}
+		elseif ($_POST['action'] == 'delete_question') {
+			if (isset($_POST['id'])) {
+				$stmt = $conn->prepare("DELETE FROM questions WHERE id like :id");
+				$stmt->bindParam(':id', $id);
+				$id = $_POST['id'];
+				$stmt->execute();
+
+				echo "The question has been deleted";
+			}
 		}
 		elseif ($_POST['action'] == 'all') {
 			$stmt = $conn->prepare("SELECT q.title, fu.url
@@ -113,6 +172,17 @@ if (isset($_POST['action'])) {
 				$stmt->execute();
 
 				echo json_encode($stmt->fetchAll(PDO::FETCH_CLASS, "questionnaire"));
+			}
+		}
+		elseif ($_POST['action'] == 'get_title_question') {
+			if (isset($_POST['id'])) {
+				$stmt = $conn->prepare("SELECT q.title 
+					FROM firsturl fu
+					INNER JOIN questionnaires q ON fu.id = q.key_first_url
+					WHERE fu.id =".$_POST['id']);
+				$stmt->execute();
+
+				echo json_encode($stmt->fetchAll(PDO::FETCH_COLUMN, 0));	
 			}
 		}
 	}
